Implement method `clifford` for all `AbstractMeasurement`
